1-lab-basicDispatchConsumeEvent/source/cdk/requirements.txt
aws-cdk.core==1.70.0
aws-cdk.aws-lambda==1.70.0
aws-cdk.aws-iam==1.70.0
aws-cdk.aws-events==1.70.0
aws-cdk.aws-events-targets==1.70.0

